feat(transfers): persistent background uploads and downloads across navigation
Lifts upload and download state into a root-level TransferProvider so progress
continues regardless of which route the user is on. Uploads targeting the current
folder still appear inline in the file list; uploads for other folders and all
zip-download progress surface in a floating panel (bottom-right, portal to body).

- transfer-context.tsx: owns uploadingFiles[], activeDownload, XHR/chunked upload
  logic, download poll, localStorage resume — exposes currentFilesViewFolderId so
  panel can scope inline vs. floating correctly
- transfer-panel.tsx: portal-rendered floating panel with upload rows + download row;
  suppresses uploads that belong to the currently visible folder
- transfer-root.tsx: client wrapper placed in root layout (no force-dynamic) to
  guarantee provider never remounts across workspace navigations
- files-view.tsx: delegates all upload/download state to context; registers
  fileInputRef + folderId so Upload button works from any route; adds ?upload=1
  auto-open; re-scans resumable sessions on listing change to prevent stale ghost
  rows after completed chunked uploads; filters ghosts by existing file names
- download-progress.tsx: stripped to re-export DownloadProgressState type only
- globals.css: transfer panel styles (.transfer-panel-*)
