+++ apps/web/app/globals.css
@@ -2569,6 +2569,158 @@ h3 {
   color: var(--color-success, hsl(142 76% 36%));
 }
 
+/* ---- Transfer panel (persistent background uploads + downloads) ---- */
+
+.transfer-panel {
+  position: fixed;
+  bottom: 16px;
+  right: 16px;
+  z-index: 50;
+  min-width: 300px;
+  max-width: 380px;
+  background: var(--card);
+  border: 1px solid var(--border);
+  border-radius: 8px;
+  box-shadow: 0 4px 16px color-mix(in oklab, var(--foreground) 12%, transparent);
+  overflow: hidden;
+}
+
+.transfer-panel-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 12px;
+  gap: 8px;
+}
+
+.transfer-panel-title {
+  font-size: 13px;
+  font-weight: 500;
+  color: var(--foreground);
+  flex: 1;
+}
+
+.transfer-panel-header-actions {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.transfer-panel-icon-btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 22px;
+  height: 22px;
+  border: none;
+  background: transparent;
+  color: var(--muted-foreground);
+  border-radius: 4px;
+  cursor: pointer;
+  padding: 0;
+}
+
+.transfer-panel-icon-btn:hover {
+  background: color-mix(in oklab, var(--foreground) 8%, transparent);
+  color: var(--foreground);
+}
+
+.transfer-panel-body {
+  border-top: 1px solid var(--border);
+  max-height: 320px;
+  overflow-y: auto;
+}
+
+.transfer-panel-row {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 8px 12px;
+  border-bottom: 1px solid color-mix(in oklab, var(--border) 60%, transparent);
+}
+
+.transfer-panel-row:last-child {
+  border-bottom: none;
+}
+
+.transfer-panel-row-top {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  min-width: 0;
+}
+
+.transfer-panel-row-icon {
+  flex-shrink: 0;
+  color: var(--muted-foreground);
+}
+
+.transfer-panel-row-icon--success {
+  flex-shrink: 0;
+  color: var(--color-success, hsl(142 76% 36%));
+}
+
+.transfer-panel-row-icon--spin {
+  flex-shrink: 0;
+  color: var(--muted-foreground);
+  animation: spin 1s linear infinite;
+}
+
+.transfer-panel-row-name {
+  flex: 1;
+  font-size: 12px;
+  color: var(--foreground);
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.transfer-panel-row-status {
+  font-size: 11px;
+  color: var(--muted-foreground);
+  white-space: nowrap;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  flex-shrink: 0;
+}
+
+.transfer-panel-row-status.is-error {
+  color: var(--destructive);
+}
+
+.transfer-panel-row-action {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  background: transparent;
+  color: var(--muted-foreground);
+  cursor: pointer;
+  padding: 1px 3px;
+  border-radius: 3px;
+  font-size: 11px;
+}
+
+.transfer-panel-row-action:hover {
+  background: color-mix(in oklab, var(--foreground) 8%, transparent);
+  color: var(--foreground);
+}
+
+.transfer-panel-row-track {
+  height: 3px;
+  background: color-mix(in oklab, var(--foreground) 10%, transparent);
+  border-radius: 2px;
+  overflow: hidden;
+}
+
+.transfer-panel-row-fill {
+  height: 100%;
+  background: var(--primary);
+  border-radius: 2px;
+  transition: width 0.3s ease;
+}
+
 /* ---- List container ---- */
 
 .explorer-list {
